mini refactor on content script
