+++ entrypoints/content/index.ts
@@ -2,16 +2,17 @@ import { defineContentScript } from '#imports'
 import { getConfig, setupConfigListener } from '@/services/configService'
 import { DOMObserver } from '@/services/domObserver'
 import { TextProcessor } from '@/services/textProcessor'
-import type { Names, UserSettings, ParsingStatus } from '@/utils/types'
+import type { Names, UserSettings } from '@/utils/types'
 import {
   blockContent,
   unblockContent,
   waitUntilDOMReady,
   createReplacementsMap,
   setStyle,
+  shouldParseSite,
 } from './utils'
 import { debugLog, haveNamesChanged, registerKeyboardShortcut } from '@/utils'
-import { storage } from '#imports'
+import { updateParsingStatus } from '@/services/parsingStatusService'
 
 let currentObserver: DOMObserver | null = null
 let previousEnabled: boolean | undefined = undefined
@@ -34,48 +35,36 @@ function cleanupAndReset() {
   previousTheme = undefined
 }
 
-async function updateParsingStatus(status: Omit<ParsingStatus, 'site' | 'timestamp'>) {
-  await storage.setItem('local:parsingStatus', {
-    ...status,
-    site: window.location.hostname,
-    timestamp: Date.now(),
-  } as ParsingStatus)
-}
-
 async function configureAndRunProcessor({ config }: { config: UserSettings }): Promise<void> {
   // Check if site should be parsed
   const shouldParse = shouldParseSite({ config })
 
-  // Handle any transition to disabled state (either by extension disable or blocklist/whitelist)
-  if ((!config.enabled || !shouldParse) && previousEnabled) {
+  // Handle any transition to disabled state (either by extension disable or blocklist/allowlist)
+  if (!config.enabled && previousEnabled) {
     cleanupAndReset()
-    if (!config.enabled) {
-      await updateParsingStatus({ isParsing: false, reason: 'disabled' })
-      await debugLog('extension disabled')
-    }
-    else if (!shouldParse) {
-      await updateParsingStatus({ isParsing: false, reason: 'blocked' })
-      await debugLog('not parsing site, blocklist or allowlist is set')
-    }
+    await updateParsingStatus({
+      status: { isParsing: false, reason: 'disabled' },
+      hostname: window.location.hostname,
+    })
+    await debugLog('extension disabled')
     return
   }
 
-  // Skip if disabled or site is blocked/not whitelisted
-  if (!config.enabled || !shouldParse) {
-    if (!config.enabled) {
-      await updateParsingStatus({ isParsing: false, reason: 'disabled' })
-    }
-    else {
-      await updateParsingStatus({ isParsing: false, reason: 'blocked' })
-    }
-    previousEnabled = false
-    previousNames = undefined
-    previousTheme = undefined
+  if (!shouldParse && previousEnabled) {
+    cleanupAndReset()
+    await updateParsingStatus({
+      status: { isParsing: false, reason: 'blocked' },
+      hostname: window.location.hostname,
+    })
+    await debugLog('not parsing site, blocklist or allowlist is set')
     return
   }
 
   // If we reach here, parsing is enabled
-  await updateParsingStatus({ isParsing: true, reason: 'enabled' })
+  await updateParsingStatus({
+    status: { isParsing: true, reason: 'enabled' },
+    hostname: window.location.hostname,
+  })
 
   // Check if names, theme, or highlightReplacedNames have changed
   const namesChanged = previousEnabled && haveNamesChanged(previousNames, config.names)
@@ -145,34 +134,6 @@ async function configureAndRunProcessor({ config }: { config: UserSettings }): P
   previousNames = config.names
   previousTheme = config.theme
   previousHighlight = config.highlightReplacedNames
-}
-
-function getMostSpecificMatch(list: string[], site: string): string | null {
-  // Return the longest matching prefix
-  const matches = list.filter(entry => site.startsWith(entry))
-  if (matches.length === 0) return null
-  return matches.reduce((a, b) => (b.length > a.length ? b : a))
-}
-
-// check config global block list and white list if the parsing should continue
-function shouldParseSite({ config }: { config: UserSettings }) {
-  const { hostname, pathname } = window.location
-  const fullUrl = `${hostname.replace(/^www\./, '')}${pathname}`
-  const allowMatch = getMostSpecificMatch(config.allowlist, fullUrl)
-  const blockMatch = getMostSpecificMatch(config.blocklist, fullUrl)
-
-  if (config.defaultAllowMode) {
-    if (!blockMatch) return true
-    if (!allowMatch) return false
-
-    return allowMatch.length >= blockMatch.length
-  }
-  else {
-    if (!allowMatch) return false
-    if (!blockMatch) return true
-
-    return allowMatch.length >= blockMatch.length
-  }
 }
 
 export default defineContentScript({
