+++ entrypoints/content/utils.ts
@@ -91,3 +91,70 @@ export function setStyle({
 export function getDataKey(attr: string): string {
   return kebabToCamel(`deadname-${attr}`)
 }
+
+/**
+ * Finds the most specific (longest) matching prefix from a list of strings against a target site.
+ *
+ * This function is used to determine the best matching entry from allowlist/blocklist configurations
+ * by comparing URL patterns. It returns the entry with the longest matching prefix, which represents
+ * the most specific match.
+ *
+ * @param list - Array of string patterns to match against (e.g., allowlist/blocklist entries)
+ * @param site - The target site string to match (typically a URL or path)
+ * @returns The longest matching prefix from the list, or null if no matches are found
+ *
+ * @example
+ * ```typescript
+ * const allowlist = ['example.com', 'example.com/admin', 'example.com/user']
+ * const site = 'example.com/admin/dashboard'
+ * const match = getMostSpecificMatch(allowlist, site)
+ * // Returns: 'example.com/admin' (longest matching prefix)
+ * ```
+ */
+function getMostSpecificMatch(list: string[], site: string): string | null {
+  // Return the longest matching prefix
+  const matches = list.filter(entry => site.startsWith(entry))
+  if (matches.length === 0) return null
+  return matches.reduce((a, b) => (b.length > a.length ? b : a))
+}
+
+/**
+ * Determines whether name replacement should be performed on the current site based on allowlist/blocklist configuration.
+ *
+ * This function evaluates the current URL against the user's allowlist and blocklist settings to decide
+ * if the deadname remover should process the page. It uses a priority system where more specific matches
+ * (longer prefixes) take precedence over less specific ones.
+ *
+ * @param obj.config - Configuration object containing user settings, including allowlist, blocklist, and defaultAllowMode
+ * @returns true if the site should be parsed for name replacement, false otherwise
+ *
+ * @example
+ * ```typescript
+ * const config = {
+ *   allowlist: ['example.com/admin'],
+ *   blocklist: ['example.com'],
+ *   defaultAllowMode: true
+ * }
+ * const shouldParse = shouldParseSite({ config })
+ * // Returns: true (allowlist match is more specific than blocklist match)
+ * ```
+ */
+export function shouldParseSite({ config }: { config: UserSettings }) {
+  const { hostname, pathname } = window.location
+  const fullUrl = `${hostname.replace(/^www\./, '')}${pathname}`
+  const allowMatch = getMostSpecificMatch(config.allowlist, fullUrl)
+  const blockMatch = getMostSpecificMatch(config.blocklist, fullUrl)
+
+  if (config.defaultAllowMode) {
+    if (!blockMatch) return true
+    if (!allowMatch) return false
+
+    return allowMatch.length >= blockMatch.length
+  }
+  else {
+    if (!allowMatch) return false
+    if (!blockMatch) return true
+
+    return allowMatch.length >= blockMatch.length
+  }
+}
